+++ backend/app.js
@@ -24,12 +24,12 @@ app.use("/api/order", order);
 app.use("/api/payment", payment);
 
 // deployment
-__dirname = path.resolve();
+const rootDir = path.resolve();
 if (process.env.NODE_ENV === "production") {
-  app.use(express.static(path.join(__dirname, "/frontend/build")));
+  app.use(express.static(path.join(rootDir, "/frontend/build")));
 
   app.get("*", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "frontend", "build", "index.html"));
+    res.sendFile(path.resolve(rootDir, "frontend", "build", "index.html"));
   });
 } else {
   app.get("/", (req, res) => {