+++ backend/middlewares/helpers/dbConnection.js
@@ -11,7 +11,7 @@ module.exports = () => {
       useFindAndModify: false,
     })
     .then(() => console.log("DB Connected"))
-    .catch((err) => {
+    .catch((_err) => {
       console.error(
         "Failed to connect to the database on startup - retrying in 5 sec",
       );