+++ backend/models/ManualOrder.js
@@ -1,5 +1,4 @@
 const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
 const manualOrderSchema = mongoose.Schema(
   {
     productName: {