+++ backend/models/RefereshToken.js
@@ -1,5 +1,4 @@
 const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
 const tokenSchema = mongoose.Schema({
   //we need user ref in order to create accessToken of that user only
   // user: {